+++ gateway/gateway-controller/pkg/utils/mcp_transformer_test.go
@@ -21,6 +21,8 @@ package utils
 import (
 	"testing"
 
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
 	api "github.com/wso2/api-platform/gateway/gateway-controller/pkg/api/management"
 	"github.com/wso2/api-platform/gateway/gateway-controller/pkg/constants"
 )
@@ -477,4 +479,172 @@ func TestGetParamsOfPolicy_MCP(t *testing.T) {
 	if _, ok := request["headers"]; !ok {
 		t.Error("Expected request.headers in params")
 	}
+}
+
+// ============================================================================
+// Resilience route-mapping tests
+// ============================================================================
+
+// mcpOpResilience returns the resilience block attached to the operation matching method+path,
+// or nil if the operation is absent or carries no resilience.
+func mcpOpResilience(ops []api.Operation, method api.OperationMethod, path string) *api.Resilience {
+	for _, op := range ops {
+		if op.Method == method && op.Path == path {
+			return op.Resilience
+		}
+	}
+	return nil
+}
+
+// When no resilience is set, the MCP forwarding routes (GET/POST/DELETE /mcp) default the route
+// timeout to "0s" (disabled) — MCP is streaming, so a finite total cap would sever a live stream.
+// The local-response routes (OPTIONS, PRM) must carry no resilience. See mcp-timeout-divergence.md.
+func TestMCPTransform_Resilience_DefaultsRouteTimeoutDisabled(t *testing.T) {
+	context := "/everything"
+	latest := LATEST_SUPPORTED_MCP_SPEC_VERSION
+	in := &api.MCPProxyConfiguration{
+		Spec: api.MCPProxyConfigData{
+			DisplayName: "everything",
+			Version:     "v1.0",
+			Context:     &context,
+			SpecVersion: &latest,
+			Upstream:    api.MCPProxyConfigData_Upstream{Url: stringPtr("http://backend:8080")},
+			// A cors policy makes the transformer emit the OPTIONS routes so we can assert they are skipped.
+			Policies: &[]api.Policy{{Name: "cors", Version: "v1"}},
+		},
+	}
+
+	res, err := (&MCPTransformer{}).Transform(in, &api.RestAPI{})
+	require.NoError(t, err)
+	ops := res.Spec.Operations
+
+	for _, m := range []api.OperationMethod{api.OperationMethodGET, api.OperationMethodPOST, api.OperationMethodDELETE} {
+		r := mcpOpResilience(ops, m, constants.MCP_RESOURCE_PATH)
+		require.NotNil(t, r, "forwarding %s %s should carry resilience", m, constants.MCP_RESOURCE_PATH)
+		require.NotNil(t, r.Timeout)
+		assert.Equal(t, "0s", *r.Timeout, "route timeout should default to disabled for MCP")
+		assert.Nil(t, r.IdleTimeout, "idle timeout should be left to the global default when unset")
+	}
+
+	// Local-response routes must not carry a route timeout.
+	assert.Nil(t, mcpOpResilience(ops, api.OperationMethodOPTIONS, constants.MCP_RESOURCE_PATH), "OPTIONS /mcp is a local reply")
+	assert.Nil(t, mcpOpResilience(ops, api.OperationMethodGET, constants.MCP_PRM_RESOURCE_PATH), "PRM route is a local reply")
+	assert.Nil(t, mcpOpResilience(ops, api.OperationMethodOPTIONS, constants.MCP_PRM_RESOURCE_PATH), "OPTIONS PRM is a local reply")
+}
+
+// An explicit resilience block overrides the disabled default and its idleTimeout is carried through,
+// on the forwarding routes only.
+func TestMCPTransform_Resilience_UserOverride(t *testing.T) {
+	context := "/everything"
+	latest := LATEST_SUPPORTED_MCP_SPEC_VERSION
+	in := &api.MCPProxyConfiguration{
+		Spec: api.MCPProxyConfigData{
+			DisplayName: "everything",
+			Version:     "v1.0",
+			Context:     &context,
+			SpecVersion: &latest,
+			Upstream:    api.MCPProxyConfigData_Upstream{Url: stringPtr("http://backend:8080")},
+			Resilience:  &api.Resilience{Timeout: stringPtr("30s"), IdleTimeout: stringPtr("120s")},
+		},
+	}
+
+	res, err := (&MCPTransformer{}).Transform(in, &api.RestAPI{})
+	require.NoError(t, err)
+	ops := res.Spec.Operations
+
+	r := mcpOpResilience(ops, api.OperationMethodGET, constants.MCP_RESOURCE_PATH)
+	require.NotNil(t, r)
+	require.NotNil(t, r.Timeout)
+	assert.Equal(t, "30s", *r.Timeout)
+	require.NotNil(t, r.IdleTimeout)
+	assert.Equal(t, "120s", *r.IdleTimeout)
+
+	// The PRM route (present for spec >= 2025-06-18) is still skipped.
+	assert.Nil(t, mcpOpResilience(ops, api.OperationMethodGET, constants.MCP_PRM_RESOURCE_PATH))
+}
+
+// A user timeout with no idle override: timeout is the user value, idle stays unset (global default).
+func TestMCPTransform_Resilience_TimeoutOnly(t *testing.T) {
+	context := "/everything"
+	latest := LATEST_SUPPORTED_MCP_SPEC_VERSION
+	in := &api.MCPProxyConfiguration{
+		Spec: api.MCPProxyConfigData{
+			DisplayName: "everything",
+			Version:     "v1.0",
+			Context:     &context,
+			SpecVersion: &latest,
+			Upstream:    api.MCPProxyConfigData_Upstream{Url: stringPtr("http://backend:8080")},
+			Resilience:  &api.Resilience{Timeout: stringPtr("45s")},
+		},
+	}
+
+	res, err := (&MCPTransformer{}).Transform(in, &api.RestAPI{})
+	require.NoError(t, err)
+
+	r := mcpOpResilience(res.Spec.Operations, api.OperationMethodPOST, constants.MCP_RESOURCE_PATH)
+	require.NotNil(t, r)
+	require.NotNil(t, r.Timeout)
+	assert.Equal(t, "45s", *r.Timeout)
+	assert.Nil(t, r.IdleTimeout)
+}
+
+// ============================================================================
+// Upstream ref / connect-timeout threading tests
+// ============================================================================
+
+// The MCP converter maps an upstream `ref` onto the derived RestAPI and threads the
+// upstreamDefinitions through, so the per-upstream connect timeout resolves like it does for RestApi.
+func TestMCPTransform_UpstreamRef_ThreadsDefinitions(t *testing.T) {
+	context := "/everything"
+	latest := LATEST_SUPPORTED_MCP_SPEC_VERSION
+	defs := []api.UpstreamDefinition{{
+		Name:    "mcp-backend",
+		Timeout: &api.UpstreamTimeout{Connect: stringPtr("6s")},
+	}}
+	in := &api.MCPProxyConfiguration{
+		Spec: api.MCPProxyConfigData{
+			DisplayName:         "everything",
+			Version:             "v1.0",
+			Context:             &context,
+			SpecVersion:         &latest,
+			UpstreamDefinitions: &defs,
+			Upstream:            api.MCPProxyConfigData_Upstream{Ref: stringPtr("mcp-backend")},
+		},
+	}
+
+	var out api.RestAPI
+	res, err := (&MCPTransformer{}).Transform(in, &out)
+	require.NoError(t, err)
+
+	require.NotNil(t, res.Spec.Upstream.Main.Ref)
+	assert.Equal(t, "mcp-backend", *res.Spec.Upstream.Main.Ref)
+	assert.Nil(t, res.Spec.Upstream.Main.Url)
+
+	require.NotNil(t, res.Spec.UpstreamDefinitions)
+	require.Len(t, *res.Spec.UpstreamDefinitions, 1)
+	require.NotNil(t, (*res.Spec.UpstreamDefinitions)[0].Timeout)
+	assert.Equal(t, "6s", *(*res.Spec.UpstreamDefinitions)[0].Timeout.Connect)
+}
+
+func TestMCPTransform_UpstreamUrl_Unchanged(t *testing.T) {
+	context := "/everything"
+	latest := LATEST_SUPPORTED_MCP_SPEC_VERSION
+	in := &api.MCPProxyConfiguration{
+		Spec: api.MCPProxyConfigData{
+			DisplayName: "everything",
+			Version:     "v1.0",
+			Context:     &context,
+			SpecVersion: &latest,
+			Upstream:    api.MCPProxyConfigData_Upstream{Url: stringPtr("http://backend:8080")},
+		},
+	}
+
+	var out api.RestAPI
+	res, err := (&MCPTransformer{}).Transform(in, &out)
+	require.NoError(t, err)
+
+	require.NotNil(t, res.Spec.Upstream.Main.Url)
+	assert.Equal(t, "http://backend:8080", *res.Spec.Upstream.Main.Url)
+	assert.Nil(t, res.Spec.Upstream.Main.Ref)
+	assert.Nil(t, res.Spec.UpstreamDefinitions)
 }
